Add support MS Visual Studio 2008
For using in VS2008:
#include <stdint.h>
#define ASMJIT_STATIC
#define ASMJIT_SUPPRESS_STD_TYPES
#include "asmjit/asmjit.h"

+++ src/asmjit/base/operand.h
@@ -57,7 +57,7 @@ struct Operand_ {
     // |XXXXXXXX|........|........|........|
     kSignatureSizeShift         = 24,
     kSignatureSizeBits          = 0xFFU,
-    kSignatureSizeMask          = kSignatureSizeBits << kSignatureSizeShift,
+    kSignatureSizeMask          = (uint32_t)kSignatureSizeBits << kSignatureSizeShift,
 
     // Register type (5 bits).
     // |........|........|........|XXXXX...|
